COG-332: CognitoUtils i moved to util package
